fix: sandbox repo code page and execution details - copy / download / step nav
