+++ apps/gitness/src/utils/common-utils.ts
@@ -1,0 +1,22 @@
+import { LivelogLine } from '@harnessio/code-service-client'
+
+export const createAndDownloadBlob = (output: string, fileName: string = 'logs') => {
+  if (!output?.length) return
+  const link = document.createElement('a')
+  const blob = new Blob([output], { type: 'application/text' })
+
+  link.download = `${fileName}.log`
+  link.href = URL.createObjectURL(blob)
+  link.target = '_blank'
+  document.body.appendChild(link)
+  link.click()
+  document.body.removeChild(link)
+}
+
+export const getLogsText = (logs: LivelogLine[]) => {
+  let output = ''
+  for (let i = 0; i < logs.length; ++i) {
+    output += logs[i].out
+  }
+  return output
+}
